requirements.txt
codegen==1.0
Django==2.2.3
django-appconf==1.0.3
django-compressor==2.3
django-sass-processor==0.7.3
django-summernote==0.8.11.4
django-tinymce==2.8.0
libsass==0.19.2
patterns==0.3
psycopg2==2.8.3
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
django-imagekit
